Move method out of managers into model
To accommodate package unit testing
