+++ lib/galaxy/managers/users.py
@@ -40,13 +40,13 @@ from galaxy.managers import (
 from galaxy.managers.base import combine_lists
 from galaxy.model import (
     Job,
-    Role,
     User,
     UserAddress,
     UserQuotaUsage,
 )
 from galaxy.model.base import transaction
 from galaxy.model.db.user import (
+    _cleanup_nonprivate_user_roles,
     get_user_by_email,
     get_user_by_username,
 )
@@ -889,26 +889,3 @@ def generate_next_available_username(session, username, model_class=User):
     while session.execute(select(model_class).where(model_class.username == f"{username}-{i}")).first():
         i += 1
     return f"{username}-{i}"
-
-
-def _cleanup_nonprivate_user_roles(session, user, private_role_id):
-    """
-    Delete UserRoleAssociations EXCEPT FOR THE PRIVATE ROLE;
-    Delete sharing roles that are associated with this user only;
-    Remove user email from sharing role names associated with multiple users.
-
-    Note: this method updates the session without flushing or committing.
-    """
-    user_roles = [ura for ura in user.roles if ura.role_id != private_role_id]
-    for user_role_assoc in user_roles:
-        role = user_role_assoc.role
-        if role.type == Role.types.SHARING:
-            if len(role.users) == 1:
-                # This role is associated with this user only, so we can delete it
-                session.delete(role)
-            elif user.email in role.name:
-                # Remove user email from sharing role's name
-                role.name = role.name.replace(user.email, "[USER PURGED]")
-                session.add(role)
-        # Delete user role association
-        session.delete(user_role_assoc)
